fix: 앱 실행 가능하도록 환경 설정
