+++ app/src/main/java/com/example/tiki_taka/MyplaceFragment.kt
@@ -1,4 +1,4 @@
-package com.example.k_lab
+package com.example.tiki_taka
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
